if states added

+++ python-object_relational_mapping/13-model_state_delete_a.py
@@ -17,8 +17,10 @@ if __name__ == "__main__":
     session = Session()
 
     states = session.query(State).filter(State.name.like("%a%")).all()
-    for state in states:
-        session.delete(state)
-    session.commit()
+    
+    if states:
+        for state in states:
+            session.delete(state)
+        session.commit()
 
     session.close()
